test(bridge): migrate readClassicDeposits e2e test to Playwright
